un bon aperçu de ce que ça pourra donner

+++ src/main/scala/projet_top/projection/ProjectedPoint.scala
@@ -6,7 +6,7 @@ case object OutOfMap extends ProjectedPoint {
     "ProjectedPoint[<out of map>]"
   }
 }
-case class OnMap(x: Int, y: Int) extends ProjectedPoint {
+case class OnMap(x: Double, y: Double) extends ProjectedPoint {
   override def toString: String = {
     s"ProjectedPoint[(x: ${this.x}, y: ${this.y})]"
   }
